query average time not work

+++ recruitment-business/src/main/java/pt/uc/dei/aor/g8/jobapp/business/persistence/IJobApplicationPersistenceService.java
@@ -41,9 +41,9 @@ public interface IJobApplicationPersistenceService {
 
 	public List <IJobApplicationProxy> listOfAllAppByPosition(IPositionProxy p);
 
-	public int averageTimeForFirstInterview(Date startDate, Date endDate);
+	public double averageTimeForFirstInterview(Date startDate, Date endDate);
 
-	public int averageTimeToHiring(Date startDate, Date endDate);
+	public double averageTimeToHiring(Date startDate, Date endDate);
 
 	//public void deleteJobApplication(IJobApplicationProxy jobApplicationProxy);
